before changing

+++ psf.py
@@ -40,33 +40,38 @@ class PSF():
         output_mask = self.ouput_dir + self.target_filename + '_psf_mask.fits'
 
         input_to_galfit(self.target_file, True, psf_regions, 32.5,
-                        output_config, output_fits, output_mask)
+                        output_config, output_fits, output_mask, 'none')
 
         d.set('region select all')
         d.set('region delete select')
-
+        if os.path.exists(output_fits):
+                os.remove(output_fits)
         subprocess.run(['/bin/bash', '-c', self.galfit_path+" "+output_config])
-        os.remove('galfit.01')
+        if os.path.exists(output_fits):
+            os.remove('galfit.01')
 
-        print("galfit run done, loading into DS9...")
-        d.set("mecube new "+output_fits)
-        d.set("scale mode minmax")
-        d.set("mode none")
-        d.set("zoom to fit")
-        d.set("cube play")
+            print("galfit run done, loading into DS9...")
+            d.set("mecube new "+output_fits)
+            d.set("scale mode minmax")
+            d.set("mode none")
+            d.set("zoom to fit")
+            d.set("cube play")
 
-        done = input('Are you satisfied? yes/enter to continue, no to restart > ')
-        if done == 'no':
-            self.write_config(d)
-        
-        self.config_file = output_config
-        self.config_output_file = output_fits
+            done = input('Are you satisfied? yes/enter to continue, no to restart > ')
+            if done == 'no':
+                self.write_config(d)
+            
+            self.config_file = output_config
+            self.config_output_file = output_fits
 
-        hdul = fits.open(output_fits)
-        data = hdul[2].data
-        fits.writeto(output_model, data, overwrite=True)
+            hdul = fits.open(output_fits)
+            data = hdul[2].data
+            fits.writeto(output_model, data, overwrite=True)
 
-        self.model_file = output_model
+            self.model_file = output_model
+
+        else:
+            print('Galfit crashed! Please try again')
 
     def visualize(self, d):
         if self.config_output_file is None and self.model_file is None:
